DF-7369 MMAnotherService now can receive snapshots
Former-commit-id: 03b3b6f13a88f94e9c3875dbe1f278616b1af07a

+++ Tools/ResourceEditor/Classes/Qt/DeviceInfo/DeviceInfo/MemProfWidget.h
@@ -76,7 +76,7 @@ public slots:
     void ConnectionEstablished(bool newConnection);
     void ConnectionLost(const DAVA::char8* message);
     void StatArrived(size_t itemCount);
-    void SnapshotArrived(size_t sizeTotal, size_t sizeRecv);
+    void SnapshotArrived(unsigned int sizeTotal, unsigned int sizeRecv);
 
     void RealtimeToggled(bool checked);
     void DiffClicked();
